Refined Main Code
made variable names have the same format (each new word starts with a capital letter), re-organised the code so it is easier to read, took out a line or two which was redundant (didn't affect the code or output). Variable names are now easier to understand and are explained by comments!